Clang tidy fixes, mainly narrow conversions

+++ src/include/coati/tree.hpp
@@ -33,7 +33,7 @@ struct node_t {
     float length;
     bool is_leaf;
     size_t parent{0};
-    std::vector<int> children;
+    std::vector<size_t> children;
 
     node_t(std::string name, float len, bool leaf = false, size_t ancestor = 0)
         : label{std::move(name)},
@@ -48,7 +48,7 @@ bool read_newick(const std::string& tree_file, std::string& content);
 int parse_newick(std::string content, tree_t& guide_tree);
 int aln_order(tree_t& tree, std::vector<std::pair<int, float>>& order_list);
 bool find_seq(const std::string& name, fasta_t& f, std::string& seq);
-bool find_node(tree_t& tree, const std::string& name, int& ID);
+bool find_node(tree_t& tree, const std::string& name, size_t& ID);
 bool reroot(tree_t& tree, const std::string& label);
 float distance_ref(const tree_t& tree, size_t ref, size_t node);
